+++ src/main/java/ui/gui/components/buttons/SoundButton.java
@@ -1,4 +1,4 @@
-package ui.gui.components;
+package ui.gui.components.buttons;
 
 import gamemusic.AudioPlayer;
 import javax.swing.*;
