+++ src/main/java/ui/gui/components/buttons/StartButton.java
@@ -1,4 +1,6 @@
-package ui.gui.components;
+package ui.gui.components.buttons;
+
+import ui.gui.components.dialogs.DifficultyDialog;
 
 import javax.swing.*;
 import java.io.IOException;
