+++ src/main/java/ui/gui/components/buttons/VolumeButton.java
@@ -1,4 +1,4 @@
-package ui.gui.components;
+package ui.gui.components.buttons;
 
 import gamemusic.AudioPlayer;
