+++ src/main/java/ui/gui/components/labels/BackgroundImageLabel.java
@@ -1,4 +1,4 @@
-package ui.gui.components;
+package ui.gui.components.labels;
 
 import client.StrainXMain;
 
@@ -6,7 +6,6 @@ import javax.imageio.ImageIO;
 import javax.swing.*;
 import java.awt.*;
 import java.awt.image.BufferedImage;
-import java.io.IOException;
 import java.net.URL;
 
 public class BackgroundImageLabel extends JLabel {
