+++ src/main/java/ui/gui/components/panels/AreaPanel.java
@@ -1,11 +1,10 @@
-package ui.gui.components;
+package ui.gui.components.panels;
 
 import gamemodel.mapengine.SubArea;
 
 import javax.swing.*;
 import javax.swing.border.LineBorder;
 import java.awt.*;
-import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 import java.util.List;
