+++ src/main/java/ui/gui/components/panels/StatusPanel.java
@@ -1,4 +1,4 @@
-package ui.gui.components;
+package ui.gui.components.panels;
 
 import gamecontrol.contents.CrewMember;
 
@@ -58,9 +58,6 @@ public class StatusPanel extends JPanel{
     public JPanel getP1() {return p1;}
     public void setP1(JPanel p1) {this.p1 = p1;}
     public JPanel getP2() {return p2;}
-    public void setP2(JPanel p2) {this.p2 = p2;}
     public JPanel getP3() {return p3;}
-    public void setP3(JPanel p3) {this.p3 = p3;}
     public JPanel getP4() {return p4;}
-    public void setP4(JPanel p4) {this.p4 = p4;}
 }
