+++ src/main/java/ui/gui/components/panels/SubareaPanel.java
@@ -1,4 +1,4 @@
-package ui.gui.components;
+package ui.gui.components.panels;
 
 import gamemodel.mapengine.SubArea;
 
@@ -58,5 +58,4 @@ public class SubareaPanel extends JPanel {
     }
 
     public SubArea getSubArea() { return subArea; }
-    public void setSubArea(SubArea subArea) { this.subArea = subArea; }
 }
