--- src/main/resources/locations.json
@@ -1,107 +1,0 @@
-[
-  {
-    "name": "Zone 19",
-    "description": "As you step into Zone 19, you are hit by the strong smell of decay and rot. The once neatly maintained military tents and equipment are now torn and scattered, and zombies wander aimlessly among the debris.",
-    "reusability": 1
-  },
-  {
-    "name": "Barracks",
-    "description": "The barracks is in a state of chaos. In the flickering light you can make out the shapes of overturned bunk beds and scattered supplies. The once-sturdy walls are now marred with bullet holes and blood stains. You hear the intermittent sounds of shuffling from the adjoining rooms.",
-    "reusability": 4
-  },
-  {
-    "name": "Armory",
-    "description": "The armory is a pilfered treasure trove of weapons and ammunition, but hopefully there's something useful here. It's clear there was a frantic struggle to arm personnel amidst the outbreak. It is a dangerous place - You see a number of figures shambling and moans can be heard coming from the shadows. Zombies lurk in the corners waiting to ambush you.",
-    "reusability": 2
-  },
-  {
-    "name": "Schrader Lab",
-    "description": "As you enter the once-famed Schrader Lab, a horrid stench strips your breath away as you hurry to cover your nose. The silence is only interrupted by the eerie creaking of metal and distant groans of the undead. The once pristine facility is now in shambles, with broken glass and overturned equipment scattered about. As you make your way deeper into the complex, you can't shake the ever-lingering feeling that you are being watched.",
-    "reusability": 1
-  },
-  {
-    "name": "Airstrip",
-    "description": "The airstrip is deserted, with abandoned planes and helicopters littering the tarmac. The control tower is still active, with flashing warning lights and the sounds of zombies echoing through the empty space.",
-    "reusability": 2
-  },
-  {
-    "name": "Hangar 1",
-    "description": "As you enter the hangar, you're greeted by the sight of rusted metal and dilapidated vehicles. Once bustling with activity, the place now lies abandoned, evidence of the frantic evacuation in the face of the zombie outbreak. Cobwebs cover the planes and helicopters, while broken machinery and scattered equipment litter the floor. It's clear that this place hasn't seen human life for quite some time.",
-    "reusability": 1
-  },
-  {
-    "name": "Hangar 2",
-    "description": "Stepping into Hangar 2, the sound of your footsteps echoes against the steel walls. The remnants of past operations scattered throughout the interior. The dusty aircraft and vehicles that once filled the space now sit motionless, serving as a haunting reminder of what once was. The musty air seems to carry the ghosts of the past, hinting at the horrors that took place here at the onset of the outbreak.",
-    "reusability": 1
-  },
-  {
-    "name": "Commissary",
-    "description": "The commissary is a mess, with overturned tables, broken chairs, and the pungent smell of rotting food. The once-full shelves are now a wreck, and zombies roam the room, searching for their next meal.",
-    "reusability": 2
-  },
-  {
-    "name": "Base Gym",
-    "description": "The musty smell of decay fills your nostrils. The once vibrant gym is now littered with torn and bloodstained workout mats, weightlifting equipment, and the remains of fallen soldiers who tried to make a stand. To your left, a zombie lies face down impaled on a barbell. The walls are pocked with bullet holes and covered in dried blood. A light flickers in the back office.",
-    "reusability": 1
-  },
-  {
-    "name": "Base Hospital",
-    "description": "Entering the ramshackle hospital, the deafening silence is broken only by the sounds of your footsteps echoing through the hallways. The floors covered in debris and medical equipment, scattered about as if in a hurry. Blood stains the walls and you can almost feel the desperation and fear that must have taken place here during the height of the initial outbreak. You tighten your grip on your weapon as you force yourself to venture deeper in hopes of securing vital medical supplies.",
-    "reusability": 2
-  },
-  {
-    "name": "Fuel Depot",
-    "description": "The fueling depot is filled with the strong smell of gasoline and oil. The once-neatly stacked fuel drums are now toppled, and zombies wander among the pumps. You wonder if there's snacks to be found in the convenience store.",
-    "reusability": 3
-  },
-  {
-    "name": "Power Station",
-    "description": "You can see the tall metal fence that surrounds the power station is partially bent and broken in several places. The entrance gate is hanging off its hinges, creaking in the wind. Inside, the once well-lit and buzzing power station is now dark and silent, except for the shuffling of zombies roaming the grounds. The disgusting smell of decay and the sound of shuffling feet sends chills down your spine.",
-    "reusability": 2
-  },
-  {
-    "name": "Building 161",
-    "description": "The door creaks open and you shine your flashlight into the darkness. The building looks like it once served as an office space, with a welcome desk to one side and a hall of cubicles extending off to the right. You hear shuffling and growling coming from the far end of the hallway.",
-    "reusability": 1
-  },
-  {
-    "name": "Base Hotel",
-    "description": "The glass doors of the lobby have been shattered and the front desk is in disarray. You step over broken furniture and make your way inside, the stale air hitting you like a wall. A distant growl echoes through the halls, alerting you of the danger that lurks within.",
-    "reusability": 3
-  },
-  {
-    "name": "Building 33",
-    "description": "Another plain old office building with clear signs that its fate was no different than any other buildings on the base. You doubt you could find much of interest here.",
-    "reusability": 1
-  },
-  {
-    "name": "Dining Hall",
-    "description": "Broken chairs and overturned tables are scattered throughout the room. Toward the back you spot a door leading to the kitchen.",
-    "reusability": 3
-  },
-  {
-    "name": "Base Cinema",
-    "description": "A complete mess, popcorn litters the floor and bullet holes and blood smears adorn the walls. You're not sure what you'll find here.",
-    "reusability": 2
-  },
-  {
-    "name": "Swanky's",
-    "description": "\"Where everybody knows your name.\" The bar is a wreck, and a zombie is mulling about mindlessly behind the counter. Maybe see what he's got on tap?",
-    "reusability": 1
-  },
-  {
-    "name": "Golf Course",
-    "description": "Hope you brought your running shoes.. the fairways have clumps of zombies standing about as if they were a new course feature. Better not attract their attention. And remember, play it as it lies!",
-    "reusability": 2
-  },
-  {
-    "name": "Subway",
-    "description": "Every service member's favorite sandwich establishment. Somehow it looks immaculate despite the zombie outbreak in the surrounding radius. A beacon of civilization and hope amidst a sea of abject terror.",
-    "reusability": 1
-  },
-  {
-    "name": "Random Base House",
-    "description": "You approach the front door with caution having no idea what to expect inside. Tightening your grip on your weapon, you gently nudge the door open and step inside.",
-    "reusability": 4
-  }
-]
